backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
google-generativeai==0.3.2
pandas==2.1.1
python-dotenv==1.0.0
Werkzeug==2.3.7
